// Frontend/SamaniCrm/src/app/dashboard/settings/app-setting/models/app-settings.ts
// Generated by https://quicktype.io

export interface AppSettingsDto {
  general: GeneralSettingDto;
  security: SecuritySettingDto;
}

export interface GeneralSettingDto {
  siteName: string;
}

export interface SecuritySettingDto {
  passwordComplexity: PasswordComplexity;
  requireCaptchaOnLogin: boolean;
  twoFactorAuthentication: TwoFactorAuthentication;
}

export interface PasswordComplexity {
  minLength: number;
  requireDigit: boolean;
  requireLowercase: boolean;
  requireSpecialChar: boolean;
  requireUppercase: boolean;
}

export interface TwoFactorAuthentication {
  enable: boolean;
  viaEmail: boolean;
  viaSMS: boolean;
}